+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.3] - 2023-05-29
 * Added default and parametrized constructors to `BWidgets::Support`
 * Fix center `BWidgets::RadialMeter`
+* Add class `BWidgets::ValueVisualizable`
+
 
 ## [1.1.2] - 2023-05-27
 * Added `Widget::setFocusText()` and `Widget::getFocusText()`